smartlife_integration/requirements.txt
# SmartLife Integration - Dependencies

# Tuya Integration
tinytuya>=1.13.0
tuya-connector-python>=1.2.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0

# Telegram Bot
python-telegram-bot>=20.7

# Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.4

# Monitoring
prometheus-client>=0.19.0
structlog>=24.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dev tools
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
